// vulpes_Front-End/src/Interfaces/cadernos/caderno6.tsx
export interface ICabecalho {
  id_horario_rega: number;
  zona_homo: string;
  n_sequencia: string;
  n_subparcela: string;
  area: number;
  cenario_cli: string;
  cultura: string;
  data_sementeira: string;
  n_contador: string;
  leitura_contador_1: string;
  data_plantacao: string;
  produca_total: string;
  obtida: string;
  id_rosto: number;
  last_update: string;
  create_date: string;
  uuid: string;
}